Fix errors in Lab6

+++ wp-starter-core/src/main/java/mk/ukim/finki/wp/model/Group.java
@@ -6,21 +6,21 @@ import java.util.Random;
  * Created by Frosina on 28.11.2016.
  */
 public class Group {
-    private Long id;
+    private Integer id;
     private String name;
     private Integer capacity;
     private Integer noExercise;
 
     public Group(){}
 
-    public Group(Long id, String name, Integer capacity, Integer noExercise) {
+    public Group(Integer id, String name, Integer capacity, Integer noExercise) {
         this.id = id;
         this.name = name;
         this.capacity = capacity;
         this.noExercise = noExercise;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -32,7 +32,7 @@ public class Group {
 
     public Integer getNoExercise() { return noExercise; }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
